refactor(chat): split the overlay handlers out of ChatOverlayLayer
ChatOverlayLayer was 526 lines, most of it moderation, clipboard and
block actions rather than sheets. The actions move to
hooks/useChatOverlayHandlers, which takes the current selection as
arguments instead of subscribing, so the overlay observable still has
exactly one reader and the component is back to being the sheet tree.

Also stops two effect cleanups reading a ref that can change under them:
hold the ref object rather than its value, which keeps "call the latest
callback on unmount" without the cleanup reading a stale node. The 7TV
runtime's forward-declared subscribe ref is filled in on commit instead
of during render.
